Implemented search for the admin-dashboard
This functionality will focus the cursor on the searchbox upon load and then
filter teachers in the list based on what name or email a user enters. When the enter key is
pressed, the searchbox is cleared.

This implementation currently assume that all teachers are rendered into the
html-response every time a user refreshes the page but this would work for a
client-based rendering as well with some modifications. However the server-
alternative should be just fine for our small use-case.