Add .env.example file
At the moment, one's required to create the .env based on information from
both chronos/wiki and development.rst. This could be an easier way to provide
the initial .env for new installations.

To use it, simply make a copy of the file. Whenever new keys are required, add
them to the .env.example together with the commit introducing the change.